Fixed bugs where 0-valued indices were being stored in the vector and arrays
M    util/SparseHashArray.java

- Fixed bug in .set() where null-valued indices were being put in the backing
  map

M    vector/SparseHashDoubleVector.java
M    vector/SparseHashIntegerVector.java

- Fixed set calls to avoid double autoboxing

M    vector/SparseHashVector.java

- Updated to set all 0-valued indices as null in the backing array, which
  ensures the array stays sparse.

+++ sspace/src/edu/ucla/sspace/vector/SparseHashDoubleVector.java
@@ -88,7 +88,7 @@ public class SparseHashDoubleVector extends SparseHashVector<Double>
     /**
      * {@inheritDoc}
      */
-    public void set(int index, double value) {
+    public void set(int index, double value) {        
         vector.set(index, value);
     }
 
@@ -96,7 +96,7 @@ public class SparseHashDoubleVector extends SparseHashVector<Double>
      * {@inheritDoc}
      */
     public void set(int index, Number value) {
-        vector.set(index, value.doubleValue());
+        vector.set(index, value);
     }
 
     /**
